+++ src/tests/users/test_user.py
@@ -3,6 +3,7 @@ from django.forms import model_to_dict
 from model_bakery import baker
 from rest_framework.test import APIClient
 
+from core.services.jwt import JwtToken
 from users.models import User
 from users.viewsets.user.service import RegisterUserService
 
@@ -83,12 +84,12 @@ class TestUserEndpoints:
 
     def test_validate(self, client_as_user: APIClient):
         user = baker.make(User)
-        token = RegisterUserService().create_validation_jwt(user)
+        token = JwtToken().encode(payload={'uuid': str(user.uuid)}, minutes=60)
         response = client_as_user[0].get(f"{self.validate_endpoint}?token={token}&uuid={str(user.uuid)}")
         assert response.status_code == 200
 
     def test_validate_forbidden(self, client_as_user: APIClient):
         user = baker.make(User, active=True, verified=True)
-        token = RegisterUserService().create_validation_jwt(user)
+        token = JwtToken().encode(payload={'uuid': str(user.uuid)}, minutes=60)
         response = client_as_user[0].get(f"{self.validate_endpoint}?token={token}&uuid={str(user.uuid)}")
         assert response.status_code == 403